+++ test/jest/controls/control.textMask.test.js
@@ -77,7 +77,7 @@ describe("control.text: mask", () => {
     el.focus();
     await h.wait(1);
     expect(el.innerHTML).toMatchInlineSnapshot(
-      `"<label for="txt1"><span><span aria-hidden="true" maskholder=""><i></i>yyyy-mm-dd</span><input placeholder=" " type="text" id="txt1" autocomplete="off" inputmode="numeric"><strong></strong></span><button clear="" tabindex="-1" aria-hidden="true" type="button"></button></label>"`
+      `"<label for="txt1"><span><span aria-hidden="true" maskholder=""><i></i>yyyy-mm-dd</span><input placeholder=" " type="text" id="txt1" autocomplete="off" inputmode="numeric"><strong></strong></span><button clear="" tabindex="-1" aria-hidden="true" type="button" style="display: none;"></button></label>"`
     );
 
     const typeText = [
@@ -210,7 +210,7 @@ describe("control.text: mask", () => {
     el.focus();
     await h.wait(1);
     expect(el.innerHTML).toMatchInlineSnapshot(
-      `"<label for="txt1"><span><span aria-hidden="true" maskholder=""><i></i>yyyy/mm/dd</span><input placeholder=" " type="text" id="txt1" autocomplete="off" inputmode="numeric"><strong></strong></span><button clear="" tabindex="-1" aria-hidden="true" type="button"></button></label>"`
+      `"<label for="txt1"><span><span aria-hidden="true" maskholder=""><i></i>yyyy/mm/dd</span><input placeholder=" " type="text" id="txt1" autocomplete="off" inputmode="numeric"><strong></strong></span><button clear="" tabindex="-1" aria-hidden="true" type="button" style="display: none;"></button></label>"`
     );
 
     const typeText = [
@@ -335,7 +335,7 @@ describe("control.text: mask", () => {
     el.focus();
     await h.wait(1);
     expect(el.innerHTML).toMatchInlineSnapshot(
-      `"<label for="txt1"><span><span aria-hidden="true" maskholder=""><i></i>xxx.xxx.xxx.xxx</span><input placeholder=" " type="text" id="txt1" autocomplete="off" inputmode="numeric"><strong></strong></span><button clear="" tabindex="-1" aria-hidden="true" type="button"></button></label>"`
+      `"<label for="txt1"><span><span aria-hidden="true" maskholder=""><i></i>xxx.xxx.xxx.xxx</span><input placeholder=" " type="text" id="txt1" autocomplete="off" inputmode="numeric"><strong></strong></span><button clear="" tabindex="-1" aria-hidden="true" type="button" style="display: none;"></button></label>"`
     );
 
     const typeText = [
@@ -582,15 +582,19 @@ describe("control.text: mask", () => {
     expect(parse("$ 5 USD", mask).leftLength).toBe(0);
 
     expect(h.getInputCursor(el.$refInput)).toBe("|");
+    expect(el.$refBtnClear?.style.display).toBe("none");
     el.$options.mask = mask;
     el.focus();
     expect(h.getInputCursor(el.$refInput)).toBe("$ |");
     await h.wait(1);
     expect(h.getInputCursor(el.$refInput)).toBe("$ |");
+    expect(el.$refBtnClear?.style.display).toBe("none");
 
     await h.userTypeText(el.$refInput, "5", { clearPrevious: false });
     await h.wait(150);
     expect(h.getInputCursor(el.$refInput)).toBe("$ 5| USD");
+    expect(el.$refBtnClear?.style.display).toBeFalsy();
+
     await h.userTypeText(el.$refInput, "2", { clearPrevious: false });
     await h.wait(150);
     expect(h.getInputCursor(el.$refInput)).toBe("$ 52| USD");